// src/calculation/enum/project_type.enum.ts
export enum ProjectTypeEnum {


    expost = 1,

    exante = 2,

    baseLineApproch_A = 3,
    baseLineApproch_B = 4,
    baseLineApproch_C = 5,

    projectApproch_A = 6,
    projectApproch_B = 7,
    projectApproch_C = 8,
    simplified = 9,
    roadpricingSimlified = 10,
    cordonPricing = 11,


}